SD-6: extract CSS and the biggest components out of App.jsx
2,216 lines to 1,376 — 38% out, past the ~600-line target in the audit. Nothing
was rewritten; everything moved verbatim.

CSS first, because that was the audit's concrete complaint: styles lived in five
template literals inside App.jsx, so component styles could not be found by
searching .css files and the RTL audit had to grep JSX for `border-left`. Now
src/styles/{base,forms,cards,timer,onboarding}.css, imported in the old
concatenation order so the cascade is unchanged.

Verified rather than assumed: extracted the same five blocks out of the previous
commit's App.jsx, concatenated them, and diffed against the new files with
comments stripped. 370 rules against 370, identical and in order. The CSS also
now ships as a real 31.8 kB stylesheet instead of being injected from JS.

One trap in that extraction worth recording: cssOnboard closes with its backtick
at the END of its last content line rather than on its own, so a line-anchored
regex ran straight past it and swallowed 100 lines of the next block. Caught by
counting lines against the known boundaries before writing anything.

Then the three leaf modals to features/plan/CourseModals.jsx, and TimerView —
347 lines, the largest single component — to features/timer/. Both are leaves:
props in, callbacks out, nothing reached from App's closure.

Shared helpers they pulled with them got proper homes rather than being
duplicated: toLocalISO and addDays to lib/dates.js (both are date helpers, and
addDays already called the other two functions there), ASSIGN_TYPES and
OTHER_ASSIGN_TYPE to lib/assignTypes.js, the exam difficulty maps to
lib/examDifficulty.js, fmtMMSS into TimerView since nothing else called it.

Lint did the work the build could not. This is plain JS with no type checking,
so Vite happily built an App.jsx referencing five identifiers that had moved
out from under it — DIFFICULTY_DAYS, DIFFICULTY_COLORS, OTHER_ASSIGN_TYPE,
addDays, fmtDateFull — and only no-undef at --max-warnings 0 caught them. Same
for toLocalISO and fmtTime on the second pass. A4's point about the absence of
type checking, demonstrated.
